pyproject.toml
[project]
name = "hotwords-lex"
version = "1.0.0"
description = "网络热词自动采集系统 - 从中英文多平台抓取热词，LLM提取，ASR词库更新"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.scripts]
hotwords-lex = "hotwords_lex.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hotwords_lex"]
